Zmiany

+++ src/main/java/com/example/projektstartowy/service/BookService.java
@@ -4,6 +4,7 @@ import com.example.projektstartowy.model.BookModel;
 import com.example.projektstartowy.repo.BookRepo;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -16,6 +17,7 @@ public class BookService {
         this.bookRepo = bookRepo;
     }
 
+    @Transactional
     public BookModel addBook(BookModel book) {
         return bookRepo.save(book);
     }
@@ -24,12 +26,18 @@ public class BookService {
         return bookRepo.findAll();
     }
 
+    @Transactional
     public BookModel updateBook(BookModel book) {
         return bookRepo.save(book);
     }
 
+    @Transactional
     public void deleteBook(Long id) {
-        bookRepo.deleteById(id);
+        if (bookRepo.existsById(id)) {
+            bookRepo.deleteById(id);
+        } else {
+            throw new BookNotFoundException("Book by id " + id + " was not found");
+        }
     }
 
     public BookModel getBookById(Long id) {
